+++ src/main/java/com/bawnorton/configurable/ap/yacl/YaclSimpleDescriptionText.java
@@ -10,6 +10,6 @@ public class YaclSimpleDescriptionText extends YaclTextElement {
 
     @Override
     public String getKey() {
-        return key;
+        return "description.%s".formatted(key);
     }
 }