+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.1
+- Prevent config screen generator from init'ing on the server
+- Add "description" to description translation key
+
 # 1.0.0
 
 Initial Release